Split gdrive utils
