+++ tests/_external/__init__.py
@@ -1,0 +1,1 @@
+"""External utility functions."""
